moving from custom to "native" C# BigInteger

+++ Phantasma.Blockchain/Contracts/ExchangeContract.cs
@@ -1,14 +1,14 @@
+using System;
+using System.Linq;
+using System.Numerics;
 using Phantasma.Core.Types;
 using Phantasma.Cryptography;
 using Phantasma.Cryptography.EdDSA;
 using Phantasma.Storage;
-using Phantasma.Numerics;
 using Phantasma.Storage.Context;
-using System;
+using Phantasma.Domain;
 using static Phantasma.Blockchain.Contracts.ExchangeOrderSide;
 using static Phantasma.Blockchain.Contracts.ExchangeOrderType;
-using Phantasma.Domain;
-using System.Linq;
 
 namespace Phantasma.Blockchain.Contracts
 {
@@ -115,7 +115,7 @@ namespace Phantasma.Blockchain.Contracts
 
         private string BuildOrderKey(ExchangeOrderSide side, string baseSymbol, string quoteSymbol) => $"{side}_{baseSymbol}_{quoteSymbol}";
 
-        public BigInteger GetMinimumQuantity(BigInteger tokenDecimals) => BigInteger.Pow(10, tokenDecimals / 2);
+        public BigInteger GetMinimumQuantity(BigInteger tokenDecimals) => BigInteger.Pow(10, ((int)tokenDecimals / 2));
         public BigInteger GetMinimumTokenQuantity(IToken token) => GetMinimumQuantity(token.Decimals);
 
         public BigInteger GetMinimumSymbolQuantity(string symbol)
